chore: separate utility functions from the crawler folder
